src/test/kotlin/com/leetcode/P394Test.kt
package com.leetcode

import org.hamcrest.MatcherAssert.assertThat
import org.hamcrest.Matchers.`is`
import org.junit.jupiter.api.Test

class P394Test {
    val p = P394()

    @Test
    fun `example 01`() {
        assertThat(p.decodeString("3[a]2[bc]"), `is`("aaabcbc"))
    }

    @Test
    fun `example 02`() {
        assertThat(p.decodeString("3[a2[c]]"), `is`("accaccacc"))
    }

    @Test
    fun `example 03`() {
        assertThat(p.decodeString("2[abc]3[cd]ef"), `is`("abcabccdcdcdef"))
    }

    @Test
    fun `example 04`() {
        assertThat(
            p.decodeString("100[leetcode]"),
            `is`("leetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcodeleetcode")
        )
    }

}
